lucli update, including fancy progress bar!
Was initially gonna do something like this for progress:
https://gist.github.com/albulescu/e61979cc852e4ee8f49c

This was quite helpful, but doesn't quite work for us.

For example, that downloads to a specific file, which we don't actually want to do here.

Instead, it turns out pb has a handy NewProxyReader method, which returns an io.Reader.
Super helpful for us, as that's exactly what update.Apply requires!

That gist was useful for testing pb, but it also revealed that you can't do a
HEAD on a GitHub release URL to get the Content-Length.

For example...

```
$ curl -I -L https://github.com/lmhd/lucli/releases/download/0.5.0/lucli.darwin.amd64

HTTP/1.1 403 Forbidden
```

So that's a shame.

Fortunately, the GitHub Releases API helps us out here, by giving us the size of each asset.